lms_30/requirement.txt
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.4
crontab==0.22.9
django-cleanup==5.1.0
django-cors-headers==3.7.0
django-crontab==0.7.1
django-widget-tweaks==1.4.8
djangorestframework==3.12.2
httpie==2.4.0
idna==2.10
Pillow==8.1.0
Pygments==2.8.0
PySocks==1.7.1
requests==2.25.1
requests-toolbelt==0.9.1
six==1.15.0
urllib3==1.26.3
whitenoise==5.2.0
